+++ frb_codegen/src/error.rs
@@ -1,4 +1,4 @@
-use crate::utils::dart_repository::dart_repository::DartDependencyMode;
+use crate::utils::dart_repository::dart_repo::DartDependencyMode;
 use thiserror::Error;
 
 pub type Result<T = ()> = std::result::Result<T, Error>;
